pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Blagovest Nedkov <bnedkov@pressindex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asyncpg = "^0.29.0"
fastapi-slim = "^0.111.0"
pyjwt = "^2.8.0"
pydantic = {extras = ["email"], version = "^2.7.1"}
uvicorn = "^0.29.0"
bcrypt = "^4.1.3"
pydantic-settings = "^2.2.1"
pytz = "^2024.1"
aioredis = ">=1.0,<2.0"
sqlalchemy = "^2.0.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
